create feature Aleksander_featureREf, changed AdsMapper, AdsService, …

+++ src/main/java/com/example/adsonline/entity/Ads.java
@@ -22,7 +22,7 @@ public class Ads {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idAds", nullable = false)
-    private Long idAds;
+    private Integer idAds;
     /**
      * Заголовок объявления
      */
